refactor: 入出力周りのいくつかの処理を`synthesizer`から移動
