+++ crates/voicevox_core/src/core.rs
@@ -1,0 +1,3 @@
+mod adjust;
+
+pub(crate) use self::adjust::{ensure_minimum_phoneme_length, pad_decoder_feature};
