+++ crates/voicevox_core/src/engine/mod.rs
@@ -1,16 +1,17 @@
 mod acoustic_feature_extractor;
 mod audio_file;
 mod full_context_label;
+mod interpret_query;
 mod kana_parser;
 mod model;
 mod mora_list;
 pub(crate) mod open_jtalk;
 
 pub(crate) use self::acoustic_feature_extractor::OjtPhoneme;
+pub(crate) use self::audio_file::to_s16le_pcm;
 pub use self::audio_file::wav_from_s16le;
-pub(crate) use self::full_context_label::{
-    extract_full_context_label, mora_to_text, FullContextLabelError,
-};
+pub(crate) use self::full_context_label::{extract_full_context_label, FullContextLabelError};
+pub(crate) use self::interpret_query::{initial_process, split_mora, DecoderFeature};
 pub(crate) use self::kana_parser::{create_kana, parse_kana, KanaParseError};
 pub use self::model::{AccentPhrase, AudioQuery, Mora};
 pub(crate) use self::mora_list::mora2text;
